Test through the full stack (including varnish and router)
Our motivation was that production was broken because the whitehall
app hadn't been registered correctly with the router, but smokey
wasn't picking this up because it was testing the apps directly
and bypassing the router.

It seems more useful to test the full stack in smokey tests.

We also made the error messages more helpful on the 'I should be
able to visit:' step.

We removed an unused step and a pending feature.