Make the whitehall feature more robust.
We would occasionally see 503 errors when Jenkins ran the Smokey builds
(preview and production).  This is because the whitehall app hadn't
started after the deploy that triggered the Smokey build.  Adding this
step should ensure that the Whitehall app is always available before we
attempt to access it through the router.