Bypass varnish by adding a cache busting string
We want to test that the router configuration is correct for each app,
but we also want to rule out the possibility that we're hitting a cached
version of a page. In this latter case there's a risk that the backend app has
failed or routing misconfigured but the test still passes because it hit
a cached page.

We can't connect directly to the router from the smokey box, so this
seems to be the next best approach.