Add a test around viewing policies.
We've currently got a bug in Whitehall that prevents the policies from
rendering.  It's caused by us not requiring a library (Addressable) in
our app code.  It's difficult/impossible to test in the Whitehall
codebase because both launchy and webmock (test libraries) require
addressable.  This was the best place I could think of to add a failing
test.

I'm navigating to a policy page through the all policies list
(/government/policies) because I can't guarantee that policies will be
consistent across production and preview.  If we continue with this
approach (navigating the site rather than directly requesting URLs) then
we might want to investigate using a library that allows us to navigate
at a slightly higher level.