Make smokey test key urls on every cache server
I used curl to do this rather than Net::HTTP as curl provides a nice
simple way to override dns resolution and I couldn't see an obvious way
to do that with Net::HTTP.

The motivation behind doing this is that we've frequently seen
intermittent problems where one of the cache servers may be
mis-configured. This testing may be redundant with respect to monitoring
systems, but this is useful until such time as those are set up.