Make the tests less noisy.
Now that we're using the cucumber progress format on CI, it just looks
odd having certain things printed to STDOUT.  I think the printing of
the host and url has outlived its usefulness as our handling/reporting
on http errors is much better.  I think the printing of the port in the
router feature is unecessary and printing the port number that failed in
the assertion message is much better.