Ensure the frontend has booted before testing the cache.
This is in an attempt to reduce the number of 504 timeouts we're
receiving during the cache test.  The previous cache test assumed that
the homepage would always be served from the cache.  My guess is that
accessing the homepage will occasionally get a cache-miss which results
in the response coming from the frontend app.  If the frontend app takes
too long to respond then the router gives us a 504.  By ensuring that
the frontend app has definitely booted, we should be able to reduce this
number of 504s, at the possible expense of getting some cache-miss
failures (i.e. legitimate test failures based on the test as it stands).