dodge rate limiting for smokey
Our smokey tests are getting rate limited. Ideally what we should do is get
nginx to whitelist the local network, but this doesn't seem to be easy without
using an evil if construct.

The Tengine clone of nginx offers a limit_req_whitelist directive for this purpose:
http://tengine.taobao.org/document/http_limit_req.html
but nginx itself doesn't seem to support it.

Therefore, we abuse the X-Forwarded-For header by setting it to a random value, making
nginx think that the smokey tests are coming from a much larger number of IPs.