Make EFG login check more general
We shouldn't have to create a lender user for smokey to use. Lenders
probably wouldn't be happy with that in production. Instead, we make
the login test a more general "am I logged in?" rather than "am I logged
in as a lender user?"

Also, assert things rather than discarding the results of expressions!