Indent properly when using govuk_setenv
The lack of indentation for the call to bundler did not make it clear
that this was run within govuk_setenv. That's caused some tail chasing
when trying to debug whether a failure of Smokey in a new environment is
related to a network connection issue or a misconfiguration.

This indents the bundle command by 4 spaces to make it clear that it is
linked to the setenv command and indents the arguments to cucumber by a
further 4 spaces to make it clear they are related to the bundle exec
command.