Add feature test for transactions explorer
Currently this is deployed on GOV.UK infrastructure. In the future it
will still be served by a URL in the www.gov.uk namespace, so it
probably makes sense to keep this test in here, even when we eventually
migrate all of /performance to be served by the performance platform
infrastructure.