Boot the right application URL for EFG
When we try to boot EFG, we were asking for efg.app_domain. When you try
that in staging you are asking for efg.production.alphagov.co.uk which
points to an external address for production and fails.

By booting efg_base_url we are actually booting the right thing
(https://www.sflg.gov.uk) which has an internal alias and will work in
Staging.

/cc @jabley