Fix trade tariff search tests
The old tests used a broken URL that redirected to a 200. Smokey's
default behaviour is to follow this URL and get the final status. So,
fix those URLs.

Secondly, even if the search returned no results these tests would pass
because of the 200 status. So update (and refactor while doing so) the
tests to look for success text on the page.