Test external link tracker
Add a test that the external link tracker is correctly redirecting a
link that exists in the search index.

Note that this should be a fairly stable test against modifications to
the set of external links held in search being modified, because even
when we remove an external link from search we don't remove it from the
external link tracker.