Remove Given the "foo" application has booted steps
These are a holdover from when we were running on passenger, and the
application workers were started on demand.  Now that we're using
Unicorn everywhere, the workers are running all the time.

This used to just send a HEAD request to the base URL for the app which
would cause passenger to spin up a worker if one wasn't already running.