Fix the smoke test for the new browse design
The new browse design includes the heading and description text inside
the anchor tag, so the existing method doesn't find the section link.

I've fixed this by defining a custom method to click on a section, 
which looks first for the <h3>, and then fetches the href from its 
parent.