[#50426159] Replace XFF/TCI with Rate-Limit-Token
Replace the X-Forwarded-For and True-Client-IP headers, which are used to
workaround rate limiting, with a Rate-Limit-Token header.

This optional token is passed in from a RATE_LIMIT_TOKEN environment
variable and will be unique to this application/environment. It will allow
us to:

- preserve the correct IP in our logs
- point Smokey at the CDN instead of origin

I've added this to the following Jenkins jobs:

- Preview_Smokey
- Staging_Smokey
- Production_Smokey

And it's added to the Icinga checks in:

- https://github.gds/gds/puppet/commit/a47d8bc4c364407c8673001039dab9ab3fe05066
- https://github.gds/gds/deployment/commit/30f5ce5f86478faae2a1a87bebfa613bb36203c2