Disable timing check for the performance platform tests
The 'elapsed time' check has been failing regularly across all environments.
This is a known issue and doesn't require immediate action, so it shouldn't
trigger an error alert in nagios.

If the page takes way too long, the nginx timeout would kick in and the return
code wouldn't be 200 (which is asserted on, so that scenario will be detected).