Update smartanswers test to step through a different answer
As can be checked on preview by visiting
/student-finance-calculator/y/2013-2014/full-time, the answer kept
redirecting to the first question, because the dates in the URl
are no longer valid as answers. Therefore the test was not testing
the flow as intended.

This picks another answer to test against that is less likely to
change the possible answers.