Tag draft-stack smokey features
This tag identifies the features to run on the draft stack.

The draft stack only contains a sub-set of applications, therefore we want to be
explicit with the features that are run by the CI server.

It's worth noting that we are currently missing smokey tests for both static and
manuals-frontend.