Add draft-only smokey features
This adjusts the organisation of the smokey features such that there are
specific ones that are only run on the draft stack.

Ideally, there wouldn't be entirely separate scenarios for testing the draft
stack, but unfortunately the existing scenarios that call the "When I visit X"
step are not compatible with user sessions as under the hood they make the
requests using RestClient and not the capybara driver. Without rewriting all the
existing scenarios to use the proper driver, it would be too mucky to make these
scenarios work on both draft and non-draft stacks.