Adjust draft steps to run in the main smokey job
these steps had been set up in such a way that there was an assumption
that smokey would be invoked with `GOVUK_WEBSITE_ROOT` set to point to
the draft environment.

Given that we've moved away from having a full duplicate stack for the
draft environment, we will now only run smoke tests in one place.

This change allows steps tagged with `@draft` to be directed towards the
draft environment urls.

I also added an extra step to some of the draft scenarios to check that
the watermark shows on the page (which is triggered by the `draft` css
class on the document body).