Only alert on the draft environment signon check
After discussion with Brad, we agreed that the check that the draft
environment is protected by signon should trigger a nagios alert,
because this functionality is unique to the draft environment and could
potentially go wrong due to a misconfiguration.

The other checks are already covered sufficiently by other monitoring
checks (app running, high 500 rate etc), so they do not need to trigger
monitoring alerts.