Add monitoring for travel advice feed
We had an incident recently where we accidentally changed the content
of travel advice feeds which caused GovDelivery to send email alerts
when there were no actual updates.

This commit tests that the GOVUK_WEBSITE_ROOT environment variable
is configured correctly on our machines. In preview and staging the
expected value is always equal to the actual value. This test is
only valuable in production where it makes sure that the website
root is set to `https://www.gov.uk`.

I've also had to add a new step definition to include the value of
the environment variable which I'm not particularly happy about but
it's the best I've come up with so far. I suppose it would be nice
if it worked with all environment variables but that's beyond the
scope of what I need right now.