Add test for crown logo link
During migration to the new production environment we had a
misconfiguration of the GOVUK_WEBSITE_ROOT environment variable which
resulted in the crown logo linking to the wrong place.

This commit adds a smoke test to parse the HTML and assert that the
crown logo is linking to the right place.