Update Tariff search feature to be more robust
Updates the assertion for Trade Tariff searches to be more robust.
This allows us to search the core functionality of the page without
coupling ourselves to specifics that may naturally chnage as the
application is iterated.

I've moved the details of the assertion from the feature to the step
defenition as the they are not important to the description of the
feature.