Update find my nearest test to make html request
This was previously making a JSON request and testing some little used functionality in Frontend, which we did not test for in the app itself. We checked in Kibana and apart from Smokey no-one else appears to be posting to JSON place routes. We've chosen to remove this functionality as it simplifies the code and do not want to be supporting legacy APIs just for the sake of it.