Allow EFG to be slower
Smokey has been failing regularly on this test. We've never noticed
before, but now we do since we've started posting the smokey failures
in slack.

This bumps the threshold for "slow" requests to EFG to 10 seconds. I
think we should possibly get rid of all benchmarking in smokey, but
this is a start.