Update Smokey tests for Design Principles
`design-principles` is no longer involved in serving the service manual – it’s entirely served by `service-manual-frontend`. Move the service manual homepage test from the design principles feature to the service manual feature.

Additionally, the search for the service manual is now handled by the global `/search` in combination with `filter_manual` so this does not need to be tested separately.