Turn on Javascript in tests
This commit makes the tests use webkit to navigate the site. It does
this instead of a homegrown solution that uses `RestClient` to fetch
pages and nokogiri to parse the result.

Having webkit makes the test visits execute javascript, which means we
will be notified if javascript is broken on the site.

To get the test suite passing we've taken a few shortcuts, which will
be handled in follow up PRs.

- the `When I request {url}` step could be removed and replaced by the
`visit` step with some work
- Steps `I am testing through the full stack` and `I force a varnish
cache miss` is not necessary anymore
- Other cleanups