Add feature tests for A/B test example page
These tests check that that users who visit the A/B example page are
assigned into both A/B buckets, and that a user who visits the page
repeatedly always sees the same version of the experiment.

Even though this just checks an example A/B test, it's still a useful
smoke test of real A/B tests, because they require the CDN to be
configured correctly to assign A/B testing cookies and convert them into
custom HTTP request headers.

Because these tests require requests to go through the CDN, they are
only enabled in production.