Stop benchmarking in smoke tests
This is controversial.

I believe we should not benchmark pages in smokey.

Firstly, because the tests are brittle and often fail. This erodes
confidence in the tests, which leads people to ignore them. When
something real breaks, we might not know. Perhaps intermittently
failing tests are worse than no tests.

Secondly, smokey is not a good tool to verify the performance of an
application. Slowdowns can have many causes. When smokey thinks
something is too slow, we learn nothing about the cause.

Lastly, developers can't reasonably act on these alerts. We've seen
dozens of smokey fails over the last months, since we started
broadcasting failures on Slack. Nothing has happened to fix the
underlying cause, probably because there's not actually a real problem
here.