Don't run benchmarking tests after deploy
The bechmarking tests often fail on jenkins because they're a bit
flaky. In https://github.com/alphagov/smokey/pull/241 it was proposed
to remove them altogether.

Because there's no other way of getting alerted for consistently poor
performing pages, we agreed that the monitoring process (which runs
smokey in a loop) will keep benchmarking. Jenkins will skip the tests,
so that the after-deploy smoke tests don't fail as much.