Test that we place users in B variant on new nav
Regardless of a user having a cookie or not, when they go to the new
education pages in collections, they should be assigned to the B
variant.

However, if a URL starts with `/education` but it's a content page, then
we should keep the user's variant.

These tests will verify that the VCL configuration is working as
expected.

This depends on https://github.com/alphagov/govuk-cdn-config/pull/35
being live.