Disable analytics before each scenario
This has the exception of 1 test case where we need to send an analytics
event, but should disable Google Analytics across all other tests.

I've also added a User-Agent to poltergeist so that we can recognise
those requests, should we need to.

Finally, we will raise an error if the Google Analytics host is somehow
in the list of hosts from the driver's network traffic list. We don't
want for this to happen and there is a cucumber tag to enable it, in
case someone needs it (e.g. used for one A/B test).