Add test for related link overrides for A/B test
We have a bunch of overrides for related links on the new
Education Navigation A/B test in place across different apps. This is
hard to test in the apps as we mock requests/responses, and we've seen
it breaking when we moved the rendering app of some of those pages.

This test should be deleted once we have a solution for related links in
the new navigation.