Rewrite scenario outline as scenario
Due to the @normal priority, this is one of the things that the
nagios_check_cache.py script checks, except, when this was changed to
an outline, the results don't appear to be included. This means that
the script breaks.

To fix this, change it back to an scenario. There is another example
of a scenario outline, but because it lacks a priority, it doesn't
break the script.