Add tests for the public API
This commit adds some tests to ensure the public API is available and returning JSON. It tests the search API, content API and the various whitehall APIs. It does not test whether the data returned by the APIs is correct.

It also moves the step definition for testing for JSON responses to a more generic file since it is now used in mutiple tests.