Instantiate Poltergeist driver with url_blacklist
Instantiating the driver with a default blacklist feels a bit simpler
than setting it in the `Before` block for scenarios not tagged with
`@withanalytics`.

The blacklist applied to scenarios tagged with `@withanalytics` still
takes priority over this default. We can observe this by adding the
`@withanalytics` tag to a scenario and observing the debug output:

    # Default blacklist defined
    {"id":"0d1a4486-73b1-49ea-a519-46418fc80c0a","name":"set_url_blacklist","args":["www.google-analytics.com"]}
    {"command_id":"0d1a4486-73b1-49ea-a519-46418fc80c0a","response":true}

    # Blacklist overwritten by `Around('@withanalytics')` block
    {"id":"ac06567c-9963-42cc-bd1f-5d9d807828fc","name":"set_url_blacklist","args":[]}
    {"command_id":"ac06567c-9963-42cc-bd1f-5d9d807828fc","response":true}

I'm confident that this change doesn't reintroduce the problem we were
seeing in issue #256. I created three scenarios and tagged the second of
the three with `@withanalytics`. I enabled Poltergeist debugging and
observed that the `set_url_blacklist` command was sent with a value of
['www.google-analytics.com'] for scenarios one and three, and with a
value of [] for scenario two.

The problem described in issue #256 was caused by the implementation of
blacklisting in PR #245. Setting the `url_blacklist` on the driver at
the point it was created meant that the `set_url_blacklist` command was
only sent once, and not before each scenario as we'd expect.