Simplify blacklist configuration for @withanalytics scenarios
We don't need to explicitly reset the `url_blacklist` as this is the
default behaviour. I've observed this by creating three scenarios,
adding the `@withanalytics` tag to the second of them and inspecting the
debug output. I can see that `url_blacklist` is set to
'www.google-analytics.com' for the 1st and 3rd scenario and is set to an
empty array for the 2nd scenario. The scenarios ran sequentially.