Remove unnecessary After block for non-Analytics requests
I believe the intention of this block is to give the user some
helpful information if we detect a request to 'www.google-analytics.com'
in a scenario that hasn't been tagged with `@withanalytics`. This will
never happen because we've blacklisted 'www.google-analytics.com' in
all scenarios except those tagged `@withanalytics`.