Check what facets look like in finder pages
This tests for issues like this:
https://docs.publishing.service.gov.uk/manual/incorrect-content-in-search-or-navigation.html#there-are-blank-options-on-finder-pages

The root cause in this case is rummager holding inconsistent data.  It knows
pages are tagged to people/organisations/topics but sometimes it's missing the
thing being tagged to.

In this case finder frontend just renders the thing with a blank title.

We could handle this better in finder frontend, but I think if the labels are ever
blank it should fail smokey, as this means the finder is not working properly.