Add search step to check for unique results
Previously some of our search tests had a "the search results should have
different titles" step which intends to check that no duplicate results are
displayed in search.  However, it is possible for two different pieces of
content to have the same title, so this step is a bit flaky.  To resolve this
issue we replace this step with "the search results should be unique", which
checks the combined search result title and description for uniqueness to
determine whether duplicate search results are appearing.