Remove some mirror features
This commit removes two features that are currently failing due to a
know SSL issue on the mirrors. This is being worked on so but in the
meantime we want Smokey to be in a green state for the rest of the tests
and to be alerted appropriately for failures in the rest of the suite.

This will be reverted once the other issue is fixed.