Allow setting of SPOOF_TARGET_DOMAIN
For the migration to AWS, we need to test the service domain works
correctly before we do the switchover. To achieve this, we want to use
Smokey as an indicator that stuff works like we expect.

Add an environment variable called `SPOOF_TARGET_DOMAIN` that allows
setting a custom domain. For our use, it will be something using the top
level "integration.govuk.digital" domain, while setting the `GOVUK_WEBSITE_ROOT`
and `EXPECTED_GOVUK_WEBSITE_ROOT` to the
`integration.publishing.service.gov.uk`.

The code will update the destination, while ensuring the Host header is
set as appropriate.