Set verify_ssl to false
Although in normal circumstances we would never wish to set this to
false, there is an issue in that we are using SNI, and as far as I can
tell the library does not allow verification of the certificate.

When DNS is migrated over, it will no longer be an issue because the
requested hostname will match the certificate.

For the interests of the migration, we wish to test the functionality of
the site under the main domain rather than explicitly testing the
certificate functionality.