Add test for assets content type
This tests for the `Content-Type` header returned by asset-manager. We
have bugs in production where these files are returned without a
`Content-Type` header, causing Internet Explorer to display the file as
plain text.

The first of these files (the .docx) has had a manual fix applied where
the Content-Type header has been set explicitly on the key in S3. The
second (the .xls) relies on nginx passing through the header from
Rails. The theory is that the bug should not happen on the first file.

The bug normally takes around a day to occur due to the day-long
`max-age` returned.