Allow setting environment from custom application
On AWS we want to run tests for the publishing and stack specific
domains. We also need to add new variables to have the tests working.

The smokey tests are called from Jenkins, as a job, via `jenkins.sh`
script, and are also running as a service in the monitoring server. In
both cases environment variables are first loaded from the environment
or from a `/etc/smokey.sh` file, and then set with `govuk_setenv default`.

This makes dificult to set up the right environment in first place and
have tests running with more than one setting, so this change allows us
to change the default `govuk_setenv` to load the environment from several
Smokey applications.