Allow whitehall-frontend to run locally
In our Smokey tests, our tests to whitehall-frontend fail because we do
not expose the frontend as a public service, so we should route
internally.

We receive these failures in Icinga:

```
2017-12-19 15:16:03: Step: [FAIL] When I visit "/government/how-government-works" on the "whitehall-frontend" application
2017-12-19 15:16:03: Error: Failed to open TCP connection to whitehall-frontend.integration.publishing.service.gov.uk:443 (getaddrinfo: Name or service not known) (SocketError)
```

There is a case that we could set Smokey to do this for several
base_urls, but I will try this on these to begin with, and then see if
it can be applied elsewhere.