Change the handling of base URLs
Previously a rather awkward list of URLs was used. This replaces that
list with using Plek, and a hash of default paths. I don't think all
default paths are used, but just include all of the ones specified
previously for now.

This uses the new external_url_for method in Plek 2.1.0 to find
external URLs for services. Some changes to the step definitions are
made here to explicitly use external URLs in some circumstances.