Fix detection if Javascript has failed
We currently test whether Javascript has failed by looking at the
wrapper for the "Is there anything wrong with this page" link. We've
replaced that feature with the Feedback component so it has stopped
working.

This replaces the detection by looking at the `js-enabled` class, which
is a better indicator of whether or not there have been Javascript
errors.

https://trello.com/c/5ZyHQaB4/139-fix-smokey-test