Add scenario to ensure draft-assets is protected by signon
I noticed that visiting draft-assets is currently redirecting us to the
internal signon URL instead of the external URL (this has been fixed in
asset-manager PR 512[1]).

Adding this scenario should help us avoid similar problems in future.

I've had to add the `@draft-assets` tag and use an `Around` filter to
set the Capybara host to use draft-assets. This is very similar to how
we're using the `@draft` tag.

I initially attempted to use a combination of `Given I am testing
"draft-assets"` and `When I request "/path"` but that uses RestClient to
make the request and I don't seem to be able to get hold of the URL that
we end up on.

[1]: https://github.com/alphagov/asset-manager/pull/512