Change the asset whitehall uses for a smokey test
The previous asset was returning a 302 which seems to be related to the
parent document it is associated with being unpublished with redirect.
I'm not 100% sure if that behavious is correct or not so raised it with
the asset-management team.

I have though switched this to using a asset of a document that is
published which cuts out a level of ambiguity.