Fix failing Whitehall attachments scenario
Whitehall PR 3918[1] added functionality to redirect requests for
attachments from gov.uk to the asset host. This functionality caused the
"Whitehall assets are served" scenario to fail because it was expecting
to be able to request an asset without it being redirected.

The fix is to assert that Whitehall attachments _are_ redirected from
gov.uk to the asset host, and to separately assert that Whitehall
attachments are successfully served from the asset host.

[1]: https://github.com/alphagov/whitehall/pull/3918