Fix mirror failing tests on Integration
The mirror tests are failing on Integration because
it is ignoring the @notaws tag and trying to run the
mirror tests. Now that Integration is on AWS it can
not access the mirrors on Carrenza.

By changing the tag to @notintegration the mirror
tests will not be run on Integration.