Increase the wait time for sign-in page to redirect
The smokey tests that check the sign-in pages redirect to Goverment Gateway and
Verify are failing in production. We think this is because the default 2 second
max wait time is not enough.

The value for `Capybara.default_max_wait_time` hasn't been overwritten because it
is also being used by the A/B testing tests.