Send JS errors to $stderr
This commit changes the JS error logger to send error messages to `$stderr` rather than the default of `$stdout`. This should prevent the messages from appearing at the top of the Smokey loop JSON file and breaking it. They will still appear inside the JSON as an issue so will be flagged up.