Reinstate whitehall feature status code check
The step 'I should be able to visit' doesn't assert anything
it only calls page.visit <Path> and doesn't verify a status
code. It's not designed as a final step and probably isn't
ideal for checking a list of paths. So this commit switches to
making a single request for each path and verifying the response
code.