Add new "When I request "..." from the "..." application" step
Currently, the "When I visit "..." on the "..." application" step
doesn't actually visit the application, it just makes a request. This
is relevant when you're logging in to Signon for example.

Therefore, create a new step that does the same thing, but is named
more appropriately. This will allow changing the implementation of the
"visit" step to actually visit the page.