Remove test dependency on related links for foreign travel advice pages
This commit removes the dependency on related links for the test scenario _Check a country page loads correctly_. Related links are changeable pieces of content and therefore should not be directly tied to the automated tests which check our production environment is operational.

This was raised from the recent A/B tests around related links, which for the B variant of this test has removed the related link being relied upon.

Solo: @karlbaker02