Switch to webdrivers
chromedriver-helper is now deprecated and they recommend switching to
webdrivers. This is similar to what we're doing for
https://github.com/alphagov/govuk_test/pull/18. The reason Smokey
currently doesn't use govuk_test is that the Smokey testing environment
is very different to an app testing environment, however we could aim to
make govuk_test more customisable so we could switch over in the future.