Remove Whitehall publications finder from tests
This commit removes tests which are specific to the Whitehall publications finder, which is being removed as it does map directly to any of the new finders.

Tests for existing documents rendered by `government-frontend` and including the URL `/government/publications/path/to/doc` will continue to be rendered in this way, so these tests are not removed in this commit.

Trello: https://trello.com/c/3BuGZK2A