Use more accurate dataset count
It turns out there are multiple ways of counting the number of datasets
in CKAN:

- https://ckan.publishing.service.gov.uk/api/3/action/package_search
- https://ckan.publishing.service.gov.uk/api/3/search/dataset
- https://ckan.publishing.service.gov.uk/api/3/action/package_list

Each one gives a slightly different package count. This is because some
of them get their data from Solr and some of them get it from Postgres.
Solr is meant to be an accurate representation of the Postgres database,
but there are no guarantees and sometimes it can end up out of sync. By
using the Postgres count, we'll get the most accurate value.