Asserts that existing finders' frequency is 'Immediately'
The frequency overrider is open to all services to use, but was only intended to be used by the business finder initially. We want to ensure that we haven't accidentally overridden the default frequency for other finders.

Note that I'd have liked to have written a more readable line of code for the assertion, e.g.

  expect(page.find("input[name='frequency'][value='#{frequency}']")).to be_checked

...but because we hide the radio buttons with CSS (in order to style them), the input cannot be found. Nokogiri treats hidden items as 'not found', so we must explicitly say we expect 'visible: false'.