Add Java setup instructions
I was surprised to run `bundle install` and see a system popup about a missing Java SDK, so it's a good idea to be explicit that Java is required.

I also initially created an Oracle account in order to download the JDK, not knowing that an OpenJDK alternative exists (nor did I know I could `brew install` it!), so have documented the advised installation process here.